Moved inteeractive commands to new module
